server/requirements.txt
flask
python-chess
